% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{calculate_pval}
\alias{calculate_pval}
\title{calculate_pval}
\usage{
calculate_pval(chiSq, df)
}
\arguments{
\item{chiSq}{Vector of chiSq statistics - this is most applicable to GxEScanR results (e.g. a single column)}

\item{df}{ChiSq distribution degrees of freedom}
}
\value{
Returns a vector of chi-square p values
}
\description{
Takes a vector of ChiSq statistics, output a vector of p values
}
\examples{
calculate_pval(3.6, 1)
ex <- rchisq(10, df=1)
calculate_pval(ex, df=1)
}
