% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figi_functions.R
\name{create_descriptive_table}
\alias{create_descriptive_table}
\title{create_descriptive_table}
\usage{
create_descriptive_table(dat, outcome, exposure, covariates, covarCat)
}
\arguments{
\item{outcome}{Outcome variable}

\item{exposure}{Exposure variable}

\item{covarCat}{String vector of categorical variables (factors)}

\item{df}{Input data}

\item{covarAll}{String vector of covariates to include in table}
}
\value{
table1 table (use in Rmarkdown)
}
\description{
Convenience wrapper function around table1 package. Function is limited but works for simple cases like distribution of covariates by outcome. Code includes workaround for adding p values that the package author described in vignette.
}
\examples{
create_descriptive_table(gxe, outcome, aspirin, c('sex', 'age'), c('sex'))
}
