% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{create_glm_stratified_plot}
\alias{create_glm_stratified_plot}
\title{create_glm_stratified_plot}
\usage{
create_glm_stratified_plot(model, G, E, gxe)
}
\arguments{
\item{model}{GLM output with higher order (interaction) term in the format G*E}

\item{G}{Genotype dosage variable (string)}

\item{E}{Exposure variable (string)}

\item{gxe}{T/F, just to tell function if GxE or ExG (sorry)}
}
\value{
A stratified GxE plot (NOT exported as *.png)
}
\description{
Uses package 'effects' to create a stratified plot of GxE interactions. Incidentally, whenever you specify interactions in glm, ALWAYS USE G*E (otherwise, specify on 'gxe' flag)
}
\examples{
create_glm_stratified_plot(model1, "X5.12345678", "exposure", gxe = T)
}
