% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_qqplot}
\alias{create_qqplot}
\title{create_qqplot}
\usage{
create_qqplot(dat, exposure, covars, stat, df, filename_suffix = "")
}
\arguments{
\item{dat}{Input data}

\item{exposure}{Exposure variable}

\item{stat}{GxEScan results chi-square statistic to plot}

\item{df}{Degrees of Freedom}

\item{filename_suffix}{For convenience when you're creating test plots}
}
\value{
Writes a PNG file into /media/work/gwis/posthoc/exposure
}
\description{
Function creates a QQ plot and outputs a PNG file
}
\examples{
create_qqplot(x, 'aspirin', c('sex', 'age'), 'chiSqGxE', df = 1, "_test")
}
