% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_twostep_weighted_plot}
\alias{create_twostep_weighted_plot}
\title{create_twostep_weighted_plot}
\usage{
create_twostep_weighted_plot(
  dat,
  exposure,
  covars,
  sizeBin0,
  alpha,
  binsToPlot,
  statistic,
  filename_suffix = ""
)
}
\arguments{
\item{dat}{Input data}

\item{sizeBin0}{Initial bin size (seeds size of remaining bins)}

\item{alpha}{Overall alpha value (default should be 0.05)}

\item{binsToPlot}{Number of bins to include in plot}

\item{statistic}{Step 1 filtering chi-square statistic (depends on method desired)}
}
\value{
A weighted hypothesis plot (png file)
}
\description{
Create weighted hypothesis testing plot. First step is creating a list object based on the data.table object output from "format_data_twostep_data" function. Second step is actual plotting. Remember that it only plots the first 10 bins for legibility.
}
\examples{
create_twostep_weighted_plot(gxe_twostep, exposure = plot_exposure, covars = plot_covariates, sizeBin0 = 5, alpha = 0.05, binsToPlot = 10, statistic = 'chiSqG')
}
