% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{format_twostep_data_expectation}
\alias{format_twostep_data_expectation}
\title{format_twostep_data_expectation}
\usage{
format_twostep_data_expectation(dat, stats_step1, sizeBin0, alpha)
}
\arguments{
\item{dat}{Input data}

\item{stats_step1}{Step1 chiSq statistic}

\item{sizeBin0}{Size of initial bin (this value needs to be power optimized)}

\item{alpha}{Overall alpha level}
}
\value{
A DATA.TABLE (!) - mainly because the code I copied pasted used a dt. Contains bin number, bin logp threshold, normalized X axis information for plotting.
}
\description{
Function to format step 1 (arrange by p value) --- assume data is the gxe object from GxEScanR
}
\examples{
format_twostep_data_expectation(gxe, 'chiSqG', 5, 0.05)
}
