% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_main_effects.R
\name{get_counts_outcome_by_group}
\alias{get_counts_outcome_by_group}
\title{get_counts_outcome_by_group}
\usage{
get_counts_outcome_by_group(dat, outcome, group)
}
\arguments{
\item{dat}{Input data}

\item{outcome}{Outcome variable}

\item{group}{Group variable e.g. study_gxe}
}
\value{
Tibble with case/control counts by group.
}
\description{
Tabulate outcome by study/platform to inclusion in forest plots. Input data should not include studies with zero counts when tabulating exposure/outcome/study.
}
