% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{write_plot_title}
\alias{write_plot_title}
\title{write_plot_title}
\usage{
write_plot_title(stat, exposure, covars, N)
}
\arguments{
\item{stat}{GxEScan chi-square statistic (look up how to set/list parameter choices)}

\item{exposure}{Exposure variable (string)}

\item{covars}{Covariate string vector}

\item{N}{Sample size}
}
\value{
Plot title string
}
\description{
Function to create figure titles based on the results statistic plotted
}
