% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_manhattanplot}
\alias{create_manhattanplot}
\title{create_manhattanplot}
\usage{
create_manhattanplot(
  x,
  exposure,
  stat,
  df,
  annotation_file,
  output_dir,
  filename_suffix = ""
)
}
\arguments{
\item{exposure}{Exposure variable (string)}

\item{stat}{GxEScan results chi-square statistic}

\item{df}{Degrees of freedom}

\item{annotation_file}{File to be used for annotation. Should located in /media/work/tmp}

\item{filename_suffix}{For convenience when you're creating test plots}

\item{dat}{Input data}
}
\value{
Writes a .png file into location /media/work/gwis/posthoc/exposure
}
\description{
Function to create manhattanplot. Uses package EasyStrata - need to write 2 files: results flat text (with appropriate columns), and .ecf file which contains EasyStrata parameters.
}
\details{
NOTE: using hardcoded paths to store temporary results text files, and to read annotation file. Be careful!
}
\examples{
create_manhattanplot(x, 'aspirin', c('sex', 'age'), 'chiSqGxE', df = 1, "_test")
}
