% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{create_qqplot_ge}
\alias{create_qqplot_ge}
\title{create_qqplot_ge}
\usage{
create_qqplot_ge(dat, exposure, covars, stat, df, filename_suffix = "")
}
\arguments{
\item{dat}{Input data}

\item{exposure}{Exposure variable}

\item{stat}{GxEScan results chi-square statistic to plot}

\item{df}{Degrees of Freedom}

\item{filename_suffix}{For convenience when you're creating test plots}
}
\value{
Writes a .png file into location /media/work/gwis/posthoc/exposure
}
\description{
Function creates a QQ plot and outputs a PNG file. Using a separate function for GE, GE among Controls, and GE among Cases (Case-only analysis) just for cosmetic reasons and to omit lambda1000
}
\examples{
create_qqplot(x, 'aspirin', c('sex', 'age'), 'chiSqGxE', df = 1, "_test")
}
