% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{get_counts_outcome_by_group}
\alias{get_counts_outcome_by_group}
\title{get_counts_outcome_by_group}
\usage{
get_counts_outcome_by_group(dat, outcome, group)
}
\arguments{
\item{outcome}{Outcome variable}

\item{group}{Group variable (typically study_g)}

\item{df}{Input data}

\item{exposure}{Exposure variable}
}
\value{
Tibble with group counts (case/control)
}
\description{
Function to tabulate outcome variable counts by study/platform. Make sure to use clean covariate data generated prior to running GxEScanR. Data should be complete case in respect to exposure and basic covariates, should not include case only studies or studies with zero outcome cells.
}
\examples{
get_counts_by_outcome(gxe, outcome, aspirin, study_gxe)
}
