% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{get_estimates_e_by_group}
\alias{get_estimates_e_by_group}
\title{get_estimates_e_by_group}
\usage{
get_estimates_e_by_group(dat, outcome, exposure, group, ...)
}
\arguments{
\item{dat}{Input data}

\item{outcome}{Outcome variable}

\item{exposure}{Exposure variable}

\item{group}{Group variable e.g. study_gxe}

\item{...}{Adjustment covariates}
}
\value{
Tibble with exposure GLM estimates/se/stats/pval/95%CI by group
}
\description{
Fit regression for exposure main effect by group. Input data should not include studies with zero counts when tabulating exposure/outcome/study.
}
\examples{
run_glm_e_by_group(gxe, outcome, asp_ref, study_gxe, age_ref_imp, sex, PC1, PC2, PC3)
}
