% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{get_estimates_gxe_by_group}
\alias{get_estimates_gxe_by_group}
\title{get_estimates_gxe_by_group}
\usage{
get_estimates_gxe_by_group(df, outcome, exposure, group, dosage, ...)
}
\arguments{
\item{df}{Input data}

\item{outcome}{Outcome variable}

\item{exposure}{Exposure variable}

\item{group}{Group variable (typically study_g)}

\item{dosage}{SNP for interaction with E}

\item{...}{Adjustment Covariates}
}
\value{
Tibble with GLM estimates/se/stats/pval/95%CI by group
}
\description{
Function to get GLM estimates by group (e.g. study_gxe). Make sure data is properly formatted yeah (maybe add info on what that means exactly..). This is an alternate version that returns the gxe term in the summary table
}
\examples{
run_glm_gxe_by_group(gxe, outcome, asp_ref, study_gxe, X5.40252294, age_ref_imp, sex, PC1, PC2, PC3)
}
